Add support for error codes that vary across supported platforms.
Prior to this commit, all error codes were Linux-specific.
This was okay for all the error codes near the start that have
the same value across all supported platforms, but to truly
support the rest of the error codes we need them on all platforms.

Note that I had to move away from this being an `:enum` because
there is currently no Savi support for varying values for an enum name
across platforms, and that frankly doesn't make sense for most enums,
so it's better to switch to a more verbose `:numeric` implementation.

One side-effect here is that the named members are now using
lowercase names (to be compatible with how named members usually are
named, which may be enforced by the compiler in the future sometime).

For this next release, the uppercase names will be retained as
compatibility synonyms, but they will be removed in a future release.