Add C++ tests for cross-platform error code checks.
This adds a new Savi program to the repo, which can
print a C++ program that checks the mapping of error codes by names,
comparing what we have implemented to what the C++ compiler resolves.

This will let us quickly find out if any of the error codes
we have defined are at variance with what a C/C++ program would see.